+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Added the option to copy the text content while browsing a note.
+- Fixed UI bugs when displaying the root note of replies.
 - Keep track of read stories.
 
 ## [0.1 (94)] - 2023-11-17Z